requirements.txt
asgiref==3.2.8
certifi==2020.4.5.2
chardet==3.0.4
dj-database-url==0.5.0
django-autoslug==1.9.7
Django==2.2
django-js-asset==1.2.2
gunicorn==20.0.4
idna==2.9
Pillow==7.1.2
psycopg2-binary==2.8.5
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.2.0
boto3==1.14.51
django-storages==1.9.1
django-phonenumber-field==5.0.0
phonenumbers==8.12.11